+++ OpenMoney.InterviewExercise.Tests/QuoteOrchestratorFixture.cs
@@ -19,8 +19,8 @@ namespace OpenMoney.InterviewExercise.Tests
                 _homeInsuranceClientMock.Object,
                 _mortgageClientMock.Object);
 
-            const float deposit = 10_000;
-            const float houseValue = 100_000;
+            const decimal deposit = 10_000;
+            const decimal houseValue = 100_000;
             
             var request = new GetQuotesRequest
             {
@@ -47,8 +47,8 @@ namespace OpenMoney.InterviewExercise.Tests
                 _homeInsuranceClientMock.Object,
                 _mortgageClientMock.Object);
 
-            const float deposit = 10_000;
-            const float houseValue = 100_000;
+            const decimal deposit = 10_000;
+            const decimal houseValue = 100_000;
             
             var request = new GetQuotesRequest
             {
