+++ OpenMoney.InterviewExercise/Models/Quotes/HomeInsuranceQuote.cs
@@ -2,9 +2,8 @@ namespace OpenMoney.InterviewExercise.Models.Quotes
 {
     public class HomeInsuranceQuote
     {
-        public float MonthlyPayment { get; set; }
-        public float BuildingsCover { get; set; }
-        public float ContentsCover { get; set; }
-        //public int LengthInMonths { get; set; }
+        public decimal MonthlyPayment { get; set; }
+        public decimal BuildingsCover { get; set; }
+        public decimal ContentsCover { get; set; }
     }
 }
